+++ packages/uniforms-bootstrap4/src/gridClassName.ts
@@ -1,5 +1,11 @@
 import { filterDOMProps } from 'uniforms';
 
+declare module 'uniforms' {
+  interface FilterDOMProps {
+    grid: never;
+  }
+}
+
 filterDOMProps.register('grid');
 
 function gridClassNamePart(