+++ packages/uniforms-bridge-simple-schema/src/register.ts
@@ -17,9 +17,28 @@ SimpleSchema.extendOptions({
   ),
 });
 
-// There's no possibility to retrieve them at runtime
+// There's no possibility to retrieve them at runtime.
+declare module 'uniforms' {
+  interface FilterDOMProps {
+    autoValue: never;
+    blackbox: never;
+    custom: never;
+    decimal: never;
+    defaultValue: never;
+    exclusiveMax: never;
+    exclusiveMin: never;
+    max: never;
+    maxCount: never;
+    min: never;
+    minCount: never;
+    optional: never;
+    regEx: never;
+    trim: never;
+    type: never;
+  }
+}
+
 filterDOMProps.register(
-  'allowedValues',
   'autoValue',
   'blackbox',
   'custom',
@@ -27,7 +46,6 @@ filterDOMProps.register(
   'defaultValue',
   'exclusiveMax',
   'exclusiveMin',
-  'label',
   'max',
   'maxCount',
   'min',