+++ packages/uniforms/__tests__/filterDOMProps.ts
@@ -11,9 +11,16 @@ describe('joinName', () => {
   });
 
   it('removes registered props', () => {
+    // @ts-ignore: Do not register its type not to pollute it.
     filterDOMProps.register('__special__');
 
     expect(filterDOMProps({ __special__: true })).toEqual({});
+  });
+
+  it('ignores double registers', () => {
+    const { length } = filterDOMProps.registered;
+    filterDOMProps.register('value');
+    expect(filterDOMProps.registered).toHaveLength(length);
   });
 
   it('omits rest', () => {