+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 ## Next
 
+- **Added:** Improved typing of `filterDOMProps`. To properly register a new property, extend `FilterDOMProps` interface. [\#763](https://github.com/vazco/uniforms/issues/763)
+
+  ```ts
+  declare module 'uniforms' {
+    interface FilterDOMProps {
+      customPropToFilter: never;
+    }
+  }
+
+  filterDOMProps.register('customPropToFilter');
+  ```
+
 - **Fixed:** Escaping special characters. [\#753](https://github.com/vazco/uniforms/issues/753)
 
 ## [v3.0.0-alpha.5](https://github.com/vazco/uniforms/tree/v3.0.0-alpha.5) (2020-06-17)